Fix wrong return type.

+++ rts/python/scalar.py
@@ -158,13 +158,13 @@ def btoi_bool_i8(x):
   return np.int8(x)
 
 def btoi_bool_i16(x):
-  return np.int8(x)
+  return np.int16(x)
 
 def btoi_bool_i32(x):
-  return np.int8(x)
+  return np.int32(x)
 
 def btoi_bool_i64(x):
-  return np.int8(x)
+  return np.int64(x)
 
 def ftob_T_bool(x):
   return bool(x)
